Fix bug in 'GBK relatedness' algorithm.
The 'GBK relatedness' algorithm is responsible for traversing the
computation graph and identifying clusters of GBK nodes that are
"related" and can therefore be exectued as part of the same MapReduce
job (i.e. same MSCR). Fix bugs in the algorithm where it is being to
greedy in its determination of whether two GBKs are related.
Specifically:

- Relatedness via shared ParallelDo inputs should only include Arr-shaped
  nodes, not Exp-shaped nodes;
- Check for indirect GBK dependencies through existing related GBK
  clusters.

Add Specs that exercie the bugs.

Conflicts:
	src/main/scala/com/nicta/scoobi/impl/plan/Smart.scala
	src/test/scala/com/nicta/scoobi/acceptance/MultipleMscrSpec.scala